+++ source/extensions/isaacsim.util.camera_inspector/isaacsim/util/camera_inspector/extension.py
@@ -63,8 +63,8 @@ class Extension(omni.ext.IExt):
         menu_items = [
             make_menu_item_description(ext_id, EXTENSION_NAME, lambda a=weakref.proxy(self): a._menu_callback())
         ]
-        self._menu_items = [MenuItemDescription(name="Workflows", sub_menu=menu_items)]
-        add_menu_items(self._menu_items, "Isaac Utils")
+        self._menu_items = [MenuItemDescription(name="Robotics", sub_menu=menu_items)]
+        add_menu_items(self._menu_items, "Tools")
         self._all_cameras = []
         self._all_viewports = []
 
@@ -80,7 +80,7 @@ class Extension(omni.ext.IExt):
 
     def on_shutdown(self):
         self._usd_context = None
-        remove_menu_items(self._menu_items, "Isaac Utils")
+        remove_menu_items(self._menu_items, "Tools")
         if self._window:
             self._window = None
         gc.collect()