+++ source/extensions/isaacsim.util.merge_mesh/isaacsim/util/merge_mesh/extension.py
@@ -38,10 +38,11 @@ class Extension(omni.ext.IExt):
         )
         self._window.deferred_dock_in("Console", omni.ui.DockPolicy.DO_NOTHING)
         self._window.set_visibility_changed_fn(self._on_window)
-        self._menu_items = [
+        menu_entry = [
             make_menu_item_description(ext_id, EXTENSION_NAME, lambda a=weakref.proxy(self): a._menu_callback())
         ]
-        add_menu_items(self._menu_items, "Isaac Utils")
+        self._menu_items = [MenuItemDescription("Robotics", sub_menu=menu_entry)]
+        add_menu_items(self._menu_items, "Tools")
         self.models = {}
         self.parent_xform = None
 
@@ -170,6 +171,6 @@ class Extension(omni.ext.IExt):
 
     def on_shutdown(self):
         """Called when the extesion us unloaded"""
-        remove_menu_items(self._menu_items, "Isaac Utils")
+        remove_menu_items(self._menu_items, "Tools")
         self._window = None
         gc.collect()