+++ source/extensions/isaacsim.robot_setup.grasp_editor/docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [2.0.3] - 2024-12-03
+### Changed
+- Isaac Util menu to Tools->Robotics menu
 
 ## [2.0.2] - 2024-10-28
 ### Changed