+++ source/extensions/isaacsim.robot_setup.import_wizard/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## [1.0.2] - 2024-12-03
+### Changed
+- Isaac Util menu to Tools->Robotics menu, Isaac Examples -> Examples browser
+
+
 ## [1.0.1] - 2024-11-09
 ### Changed
 - Updated dependencies and imports after renaming