+++ source/extensions/isaacsim.ros2.bridge/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [2.0.3] - 2024-12-03
+### Changed
+- Isaac Util menu to Tools->Robotics menu
+
 
 ## [4.1.0] - 2024-11-27
 ### Added